Refactor: Simplify type casting and align variable naming/namespace scoping
- Removed redundant `static_cast` usage across multiple files.
- Standardized to lowercase register naming (e.g., `rsp`, `rbp`) in `MicroPrologEpilogSanitize` tests for consistency.
- Updated isolated structs to be scoped within unnamed namespaces.
- Adjusted includes and cleaned up unnecessary declarations in headers and source files.
- Aligned type declarations, reducing verbosity for compatibility with modern C++ standards.

+++ src/Compiler/Parser/Ast/Ast.cpp
@@ -143,7 +143,7 @@ size_t Ast::memStorageReserved() const
 {
     size_t result = 0;
     for (const Shard& shard : shards_)
-        result += static_cast<size_t>(shard.store.allocatedBytes());
+        result += shard.store.allocatedBytes();
     return result;
 }
 #endif
